Get copy() buffer back on stack (fixes #117)
Instead of allocating buffer for file copy in heap allocate it on stack.
Just use 32k instead of default 64k. This leaves 48k stack for other
things. Which is 3x bigger than default (80k - 64k) so should be okay to
keep #116 fixed.

Performance get back to the same level as vagga 0.4.1